requirements.txt
absl-py==0.9.0
altair==4.1.0
astor==0.8.1
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.14.42
botocore==1.17.42
branca==0.4.1
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
clean-text==0.2.1
click==7.1.2
cmdstanpy==0.4.0
colorama==0.4.3
confuse==1.3.0
convertdate==2.2.1
convokit==2.4.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
ephem==3.7.7.1
Flask==1.1.2
Flask-Cors==3.0.8
folium==0.11.0
fsspec==0.8.0
ftfy==5.8
gast==0.2.2
gcsfs==0.6.0
gensim==3.8.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
holidays==0.10.3
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
imageio==2.9.0
importlib-metadata==1.7.0
ipdb==0.13.3
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.5
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
llvmlite==0.33.0
LunarCalendar==0.0.9
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.4.0
msgpack==1.0.0
msgpack-numpy==0.4.6.1
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.0.3
numba==0.50.1
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.8.2
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
py==1.9.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.5.0b1
Pygments==2.6.1
pylint==2.5.3
PyMeeus==0.3.7
pyparsing==2.4.7
pyrsistent==0.16.0
pyspellchecker==0.5.5
pystan==2.19.1.1
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3fs==0.4.2
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
setuptools-git==1.2
six==1.15.0
smart-open==2.1.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
statsmodels==0.11.1
streamlit==0.65.2
tangled-up-in-unicode==0.0.6
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.9
validators==0.18.0
wagon-tools==1.1.2
wasabi==0.7.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yellowbrick==1.1
zipp==3.1.0
